+++ src/app/(shop)/catalogue/components/ProductCard.tsx
@@ -14,7 +14,7 @@ export const ProductCard = ({ product }: Props) => {
     >
       <div className="relative w-full aspect-[3/4] bg-gray-100">
         <Image
-          src={product.imageUrl[0] || '/placeholder.svg?height=700&width=600'}
+          src={product.imageUrl[0] || '/placeholder.webp'}
           alt={product.name}
           fill
           className="object-cover"
